added testing coverage for cpp engine

+++ fral/python/fral_ext.cpp
@@ -30,7 +30,7 @@ size_t blobSize(void *blob){
 
 NB_MODULE(fral_ext, m) {
 nb::class_<fral::FRAL>(m, "FRAL")
-.def(nb::init<const char *, size_t, size_t>(), nb::rv_policy::reference)
+.def(nb::init<const char *, size_t, int>(), nb::rv_policy::reference)
 .def(nb::init<const char *>(), nb::rv_policy::reference)
 .def("allocate", &fral::FRAL::allocate)
 .def("append", &fral::FRAL::append)
